+++ fabric/src/main/java/dev/enjarai/headpats/PetRendering.java
@@ -63,12 +63,13 @@ public class PetRendering {
 
     public static @Nullable Float getCameraRoll(PlayerEntity player, float tickDelta) {
         var petting = Headpats.PETTING_COMPONENT.get(player);
+        var finalFirstPersonSwayStrength = ModConfig.INSTANCE.firstPersonSwayStrength * MinecraftClient.getInstance().options.getDistortionEffectScale().getValue();
 
-        if (petting.pettedMultiplier > 0 && ModConfig.INSTANCE.firstPersonSwayStrength > 0) {
+        if (petting.pettedMultiplier > 0 && finalFirstPersonSwayStrength > 0) {
             var petTime = MathHelper.lerp(tickDelta, (float) petting.prevPettedTicks, (float) petting.pettedTicks);
             var multiplier = MathHelper.lerp(tickDelta, petting.prevPettedMultiplier, petting.pettedMultiplier);
 
-            return -MathHelper.sin(petTime * 0.4f) * multiplier * 0.1f * ModConfig.INSTANCE.firstPersonSwayStrength;
+            return -MathHelper.sin(petTime * 0.4f) * multiplier * 0.1f * (float) finalFirstPersonSwayStrength;
         }
 
         return null;